requirements.txt
# Python 3.11.9 required (Python 3.13+ also works)
# Note: torch is already installed with CUDA support, skipping version constraint
# torch>=2.1.0  # Install separately: pip install torch --index-url https://download.pytorch.org/whl/cu118
transformers>=4.36.0,<5.0.0
accelerate>=0.25.0,<1.0.0
safetensors>=0.4.0,<1.0.0
fastapi>=0.104.0,<1.0.0
uvicorn[standard]>=0.24.0,<1.0.0
pydantic>=2.5.0,<3.0.0
pydantic-settings>=2.1.0,<3.0.0
pyyaml>=6.0.1,<7.0.0
gradio>=4.7.0,<5.0.0
numpy>=1.24.0,<2.0.0
packaging>=23.0,<24.0
